New drawing board implementation fixes and optimization

+++ Source/RozpoznawanieZnakow/Point.h
@@ -2,9 +2,9 @@
 class Point
 {
 public:
-	Point(int x, int y);
+	Point(float x, float y);
 
-	int x;
-	int y;
+	float x;
+	float y;
 };
